refactor: 구독 토글 버튼 profile 폴더로 이동
